% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doMCMCs.R
\name{doMCMCs}
\alias{doMCMCs}
\title{run a set of MCMCs for performance comparison}
\usage{
doMCMCs(modelInfo, MCMCinfo, MCMCs = "nimble", nimbleMCMCdefs = list(),
  externalMCMCinfo = list(), metrics = c("mean", "median", "sd",
  "CI95_low", "CI95_upp", "efficiency_coda"))
}
\arguments{
\item{modelInfo}{A list of model-specification information.  For running 
MCMCs in nimble, this can include elements named \code{code} (model
 code as returned from `nimbleCode`), `data` (a list with data and/or 
 constants), and `inits` (a list of initial values).  For JAGS, WinBUGS 
 and OpenBUGS, many models can also be run from the same specification 
 since they use nearly the same model language.  For a different MCMC 
 engine, including a user-provided plugin, the modelInfo can contain 
 whatever elements are required for that engine.}

\item{MCMCinfo}{A list with fields `niter` (number of iterations), 
`thin` (thinning interval), `burnin` (number of iterations to discard 
from the beginning of the MCMC sample), `setSeed` (a numeric value 
passed to `set.seed` to set the random-number generator seed), 
and/or `monitors` (a character vector of variables to record in the 
MCMC sample.}

\item{MCMCs}{}

\item{nimbleMCMCdefs}{}

\item{externalMCMCinfo}{}

\item{metrics}{}
}
\description{
run one or more MCMC engines for one model specification, with 
timing and performance metrics calculated.
}
\details{

}
\examples{

}
\seealso{

}
