% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{metrics}
\alias{metrics}
\alias{MCMCmetric_mean}
\alias{MCMCmetric_median}
\alias{MCMCmetric_sd}
\alias{MCMCmetric_CI95}
\alias{MCMCmetric_CI95low}
\alias{MCMCmetric_CI95upp}
\alias{MCMCmetric_efficiency_coda}
\title{Built-in metrics for \code{MCMCresult} objects}
\usage{
MCMCmetric_mean(result)

MCMCmetric_median(result)

MCMCmetric_sd(result)

MCMCmetric_CI95(result)

MCMCmetric_CI95low(result)

MCMCmetric_CI95upp(result)

MCMCmetric_efficiency_coda(result)
}
\arguments{
\item{result}{An MCMCresult object, normally a list element returned by \code{\link{compareMCMCs}}}
}
\value{
A list that may contain elements named:
\itemize{
\item \code{byParameter}: A named list of vectors.  In each vector,
the elements correspond to parameters.  The list names
will become names in any outputs.
\item \code{byMCMC}: A named list of numbers.
}

It is also valid to return a list of such lists.
}
\description{
These functions are normally called via compareMCMCs or addMetric.
}
\details{
A metric is a summary of MCMC output.  The summary may include results for each parameter,
for each MCMC sample (across all parameters), and/or by arbitrary list. The last option is
not used by any built-in metrics.

The built-in metrics include:
\itemize{
\item mean : mean for each parameter
\item median : median for each parameter
\item sd : standard deviation for each parameter
\item CI95 : both ends of 95\% credible interval, a combination of CI95low and CI95upp
\item CI95low : lower end of 95\% credible interval
\item CI95upp : upper end of 95\% credible interval
\item efficiency or (synonomously) efficiency_coda : effective sample size (ESS) and efficiency (ESS / computation time)
}
}
