% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_MCMC_comparison_pages.R
\name{pageComponents}
\alias{pageComponents}
\alias{registerPageComponents}
\alias{unregisterPageComponents}
\alias{getPageComponents}
\title{Register, unregister and access page components used by
\code{make_MCMC_comparison_pages}}
\usage{
registerPageComponents(pageComponents)

unregisterPageComponents(name)

getPageComponents()
}
\arguments{
\item{pageComponents}{A named list of new page components to register}

\item{name}{Character name of a page component to unregister}
}
\description{
Register, unregister and access page components used by
\code{make_MCMC_comparison_pages}
}
\details{
A page component is an element that can be included in an
MCMC comparison page by naming it in the \code{pageComponents}
argument to \code{make_MCMC_comparison_pages}. See package
vignette for explanation page components.
}
\seealso{
\link{make_MCMC_comparison_pages}
}
