% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMCdefs_env.R
\name{registerMCMCengine}
\alias{registerMCMCengine}
\title{Register an MCMC function for use by compareMCMCs}
\usage{
registerMCMCengine(name, fun)
}
\arguments{
\item{name}{The name by which the MCMC function (or "engine") is identified
in the \code{MCMCs} argument to \code{\link{compareMCMCs}}.}

\item{fun}{The function that runs and times an MCMC.}
}
\description{
Register an MCMC function for use by compareMCMCs
}
\details{
See package vignette for information about the arguments
that will be passed to \code{fun} from \code{compareMCMCs} and the \code{MCMCresult}
object that should be returned by \code{fun}.

For more information, see \link{builtin_MCMCs}.

MCMCs from \code{nimble} are run in a different way, since there can be
multiple MCMCs for the same \code{nimble} model.  These are run by
\code{runNIMBLE}, which is not exported.
}
