% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics-addMetrics.R
\name{combineMetrics}
\alias{combineMetrics}
\title{Combine all metrics from a list of \code{MCMCresult} objects.}
\usage{
combineMetrics(results)
}
\arguments{
\item{results}{a list of \code{MCMCresult} objects}
}
\description{
This is useful for seeing results from multiple MCMC engines compactly.
}
